+++ qiskit/transpiler/passes/mapping/unroller.py
@@ -7,9 +7,6 @@
 
 """Pass for unrolling a circuit to a given basis."""
 
-import networkx as nx
-
-from qiskit.circuit import QuantumRegister, ClassicalRegister
 from qiskit.transpiler._basepasses import TransformationPass
 
 
@@ -22,7 +19,7 @@ class Unroller(TransformationPass):
     def __init__(self, basis=None):
         """
         Args:
-            basis (list[Instruction]): target basis gates to unroll to
+            basis (list[Gate]): Target basis gates to unroll to.
         """
         super().__init__()
         self.basis = basis or []
@@ -39,9 +36,6 @@ class Unroller(TransformationPass):
 
         Returns:
             DAGCircuit: output unrolled dag
-
-        Raises:
-            TranspilerError: if no decomposition rule is found for an op
         """
         # Walk through the DAG and expand each non-basis node
         for node in dag.get_gate_nodes():
@@ -55,31 +49,5 @@ class Unroller(TransformationPass):
             # TODO: allow choosing other possible decompositions
             decomposition_dag = self.run(decomposition_rules[0])  # recursively unroll gates
 
-            condition = current_node["condition"]
-            # the decomposition rule must be amended if used in a
-            # conditional context. delete the op nodes and replay
-            # them with the condition.
-            if condition:
-                decomposition_dag.add_creg(condition[0])
-                to_replay = []
-                for n_it in nx.topological_sort(decomposition_dag.multi_graph):
-                    n = decomposition_dag.multi_graph.nodes[n_it]
-                    if n["type"] == "op":
-                        n["op"].control = condition
-                        to_replay.append(n)
-                for n in decomposition_dag.get_op_nodes():
-                    decomposition_dag._remove_op_node(n)
-                for n in to_replay:
-                    decomposition_dag.apply_operation_back(n["op"], condition=condition)
-
-            # the wires for substitute_circuit_one are expected as qargs first,
-            # then cargs, then conditions
-            qwires = [w for w in decomposition_dag.wires
-                      if isinstance(w[0], QuantumRegister)]
-            cwires = [w for w in decomposition_dag.wires
-                      if isinstance(w[0], ClassicalRegister)]
-
-            dag.substitute_circuit_one(node,
-                                       decomposition_dag,
-                                       qwires + cwires)
+            dag.substitute_node_with_dag(node, decomposition_dag)
         return dag