+++ test/python/test_mapper.py
@@ -13,7 +13,7 @@ import unittest
 
 from qiskit import compile, execute
 from qiskit import QuantumRegister, ClassicalRegister, QuantumCircuit
-from qiskit import Aer
+from qiskit import BasicAer
 from qiskit.backends.models import BackendConfiguration
 from qiskit.backends.models.backendconfiguration import GateConfig
 from qiskit.qobj import Qobj
@@ -85,7 +85,7 @@ class TestMapper(QiskitTestCase):
 
     def setUp(self):
         self.seed = 42
-        self.backend = Aer.get_backend("qasm_simulator_py")
+        self.backend = BasicAer.get_backend("qasm_simulator")
 
     def test_mapper_overoptimization(self):
         """Check mapper overoptimization.