+++ test/python/test_parallel.py
@@ -9,8 +9,8 @@
 import os
 import time
 
+from qiskit.tools.parallel import parallel_map
 from qiskit import QuantumRegister, ClassicalRegister, QuantumCircuit
-from qiskit.transpiler._parallel import parallel_map
 from .common import QiskitTestCase